pyproject.toml
[project]
name = "newsmate"
version = "0.1.0"
description = "NewsMate daily news digest generator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "reportlab>=4.4.7",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "tqdm>=4.67.1",
]
